вернул рабочий код с багом на место =)

+++ uravn.cpp
@@ -2,17 +2,17 @@
 #include <math.h>
 using namespace std;
 
-double main()
+int main()
 {
 	std::puts(" please enter a,b and c for `ax^2+bx+c=0`:");
-	double a = 0;
-	double b = 0;
-	double c = 0;
-	double D = 0;
+	float a = 0;
+	float b = 0;
+	float c = 0;
+	float D = 0;
 
 	scanf_s("%f %f %f", &a, &b, &c);
 	getchar();
-	
+
 	D = (b * b) - (4 * a * c);
 
 	if (a == 0 && b == 0)
@@ -33,7 +33,7 @@ double main()
 	{
 		std::printf("Error");
 	}
-	
+
 
 	if (D > 0 && D != (b * b))
 	{
@@ -45,17 +45,17 @@ double main()
 		std::printf("solution:x1=%f\n x2=%f\n", x1, x2);
 
 	}
-	
-		if (D == 0 && a != 0 && b != 0 && c !=0 ); // здесь что-то не так, выводит всегда это почему-то
+
+	if (D == 0 && a != 0 && b != 0 && c != 0); // здесь что-то не так, выводит всегда это почему-то
 	{
 		double x3;
 		x3 = (-b) / (2 * a);
 		std::printf("\nsolution: %f\n", x3);
 	}
 
-	
+
 
 	return 0;
 
-	
+
 }
